Kill phase-named modules; rename to capability-first names
- phase3_training_execution → nlp_baseline_batch.
- phase3_finetune_pilot → finetune_pilot; phase3_finetune_batch → finetune_batch.
- phase4_attention_ablation → attention_ablation.
- Delete run_phase2_pipeline.py (deprecation stub; pipeline_data_prep is the entry).
- Update Makefile, benchmark-smoke workflow, all imports, READMEs, REPO_TOUR, CLAUDE.md.
- 428 unit + regression tests pass; mypy clean on the strict-typed module set.
